Add a ValueBase<TType> class
